modelform and save() working #9

+++ pilot/models.py
@@ -3,5 +3,5 @@ from django.db import models
 
 class PID(models.Model):
     id = models.AutoField(primary_key=True)
-    participant_id = models.PositiveIntegerField(unique=True, null=False, blank=False, default=None)
+    participant_id = models.PositiveSmallIntegerField(unique=True, null=False, blank=False, default=None)
